Stack fixes (#180)
* Call update bounds after transition (#175)

* Remove usage of ScenesReducer (#174)

See #173

* Update Example app to RN 0.61 (#177)

* Fix iOS Example project after upgrade to RN 61 (#179)

* Let UINavController control subcontroller view's frames.

This PR changes the way we've been handling yoga <> NavController layout interactions. Now we ignore frame updates coming from react for the main Screen view to allow NavController take the controll. In order to keep yoga working we now use `setSize` to pass the dimensions of the view back to yoga such that it can properly calculate layout of the views under Screen component.